refactor(api): replace Client interface with concrete struct
Phase 12b Step 9: Complete Client interface to struct conversion.

Major changes:
- Consolidated all internal/api methods into pkg/dot/client.go (~986 lines)
- Converted Client from interface to concrete struct
- Removed registration mechanism (RegisterClientImpl, newClientImpl)
- Deleted entire internal/api package (29 files, ~4400 lines)
- Updated cmd/dot to no longer import internal/api

This completes the Phase 12b refactoring vision:
- Domain types in internal/domain
- Client struct directly in pkg/dot
- No interface indirection
- No registration mechanism
- Clean, idiomatic Go architecture

All tests passing:
- 17/17 packages pass
- Zero linter errors
- CLI functional
- Zero breaking changes to public API

BREAKING CHANGE (internal only): internal/api package removed.
This only affects code that directly imported internal/api, which
should not exist since it was an internal package.
