--- internal/api/doctor_comprehensive_test.go
@@ -1,216 +1,0 @@
-package api_test
-
-import (
-	"context"
-	"path/filepath"
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-
-	"github.com/jamesainslie/dot/pkg/dot"
-)
-
-func TestDoctor_AllPathTypes(t *testing.T) {
-	cfg := testConfig(t)
-	ctx := context.Background()
-
-	require.NoError(t, cfg.FS.MkdirAll(ctx, cfg.TargetDir, 0755))
-	require.NoError(t, cfg.FS.MkdirAll(ctx, filepath.Join(cfg.PackageDir, "test"), 0755))
-
-	// Create various symlink scenarios
-	// 1. Valid absolute symlink
-	source1 := filepath.Join(cfg.PackageDir, "test", "file1")
-	require.NoError(t, cfg.FS.WriteFile(ctx, source1, []byte("test"), 0644))
-	link1 := filepath.Join(cfg.TargetDir, ".file1")
-	require.NoError(t, cfg.FS.Symlink(ctx, source1, link1))
-
-	// 2. Broken symlink (target doesn't exist)
-	link2 := filepath.Join(cfg.TargetDir, ".file2")
-	require.NoError(t, cfg.FS.Symlink(ctx, "/nonexistent", link2))
-
-	// 3. Regular file where symlink expected
-	link3 := filepath.Join(cfg.TargetDir, ".file3")
-	require.NoError(t, cfg.FS.WriteFile(ctx, link3, []byte("test"), 0644))
-
-	// Create manifest
-	manifestContent := []byte(`{
-		"version": "1.0",
-		"updated_at": "2024-01-01T00:00:00Z",
-		"packages": {
-			"test": {"name": "test", "installed_at": "2024-01-01T00:00:00Z", "link_count": 3, "links": [".file1", ".file2", ".file3"]}
-		},
-		"hashes": {}
-	}`)
-	manifestPath := filepath.Join(cfg.TargetDir, ".dot-manifest.json")
-	require.NoError(t, cfg.FS.WriteFile(ctx, manifestPath, manifestContent, 0644))
-
-	client, err := dot.NewClient(cfg)
-	require.NoError(t, err)
-
-	report, err := client.DoctorWithScan(ctx, dot.DefaultScanConfig())
-	require.NoError(t, err)
-
-	// Should detect issues for file2 (broken) and file3 (not symlink)
-	assert.Equal(t, dot.HealthErrors, report.OverallHealth)
-	assert.Equal(t, 3, report.Statistics.TotalLinks)
-	assert.GreaterOrEqual(t, len(report.Issues), 2)
-}
-
-func TestDoctor_WarningsVsErrors(t *testing.T) {
-	cfg := testConfig(t)
-	ctx := context.Background()
-
-	require.NoError(t, cfg.FS.MkdirAll(ctx, cfg.TargetDir, 0755))
-
-	// Create manifest with missing link
-	manifestContent := []byte(`{
-		"version": "1.0",
-		"updated_at": "2024-01-01T00:00:00Z",
-		"packages": {
-			"vim": {"name": "vim", "installed_at": "2024-01-01T00:00:00Z", "link_count": 1, "links": [".vimrc"]}
-		},
-		"hashes": {}
-	}`)
-	manifestPath := filepath.Join(cfg.TargetDir, ".dot-manifest.json")
-	require.NoError(t, cfg.FS.WriteFile(ctx, manifestPath, manifestContent, 0644))
-
-	client, err := dot.NewClient(cfg)
-	require.NoError(t, err)
-
-	report, err := client.DoctorWithScan(ctx, dot.DefaultScanConfig())
-	require.NoError(t, err)
-
-	// Missing link is an error
-	assert.Equal(t, dot.HealthErrors, report.OverallHealth)
-	hasError := false
-	for _, issue := range report.Issues {
-		if issue.Severity == dot.SeverityError {
-			hasError = true
-			break
-		}
-	}
-	assert.True(t, hasError)
-}
-
-func TestDoctor_HandlesMultiplePackagesCorrectly(t *testing.T) {
-	cfg := testConfig(t)
-	ctx := context.Background()
-
-	require.NoError(t, cfg.FS.MkdirAll(ctx, cfg.TargetDir, 0755))
-	require.NoError(t, cfg.FS.MkdirAll(ctx, filepath.Join(cfg.PackageDir, "vim"), 0755))
-	require.NoError(t, cfg.FS.MkdirAll(ctx, filepath.Join(cfg.PackageDir, "tmux"), 0755))
-
-	// Create valid links for tmux
-	tmuxSource := filepath.Join(cfg.PackageDir, "tmux", "conf")
-	require.NoError(t, cfg.FS.WriteFile(ctx, tmuxSource, []byte("test"), 0644))
-	tmuxLink := filepath.Join(cfg.TargetDir, ".tmux.conf")
-	require.NoError(t, cfg.FS.Symlink(ctx, tmuxSource, tmuxLink))
-
-	// vim link is missing
-
-	manifestContent := []byte(`{
-		"version": "1.0",
-		"updated_at": "2024-01-01T00:00:00Z",
-		"packages": {
-			"vim": {"name": "vim", "installed_at": "2024-01-01T00:00:00Z", "link_count": 1, "links": [".vimrc"]},
-			"tmux": {"name": "tmux", "installed_at": "2024-01-01T00:00:00Z", "link_count": 1, "links": [".tmux.conf"]}
-		},
-		"hashes": {}
-	}`)
-	manifestPath := filepath.Join(cfg.TargetDir, ".dot-manifest.json")
-	require.NoError(t, cfg.FS.WriteFile(ctx, manifestPath, manifestContent, 0644))
-
-	client, err := dot.NewClient(cfg)
-	require.NoError(t, err)
-
-	report, err := client.DoctorWithScan(ctx, dot.DefaultScanConfig())
-	require.NoError(t, err)
-
-	// Should have 2 total links, 1 broken
-	assert.Equal(t, 2, report.Statistics.TotalLinks)
-	assert.Equal(t, 1, report.Statistics.BrokenLinks)
-	assert.Equal(t, dot.HealthErrors, report.OverallHealth)
-}
-
-func TestDoctor_RelativeSymlinkResolution(t *testing.T) {
-	cfg := testConfig(t)
-	ctx := context.Background()
-
-	require.NoError(t, cfg.FS.MkdirAll(ctx, cfg.TargetDir, 0755))
-	require.NoError(t, cfg.FS.MkdirAll(ctx, filepath.Join(cfg.PackageDir, "vim"), 0755))
-
-	// Create source
-	sourcePath := filepath.Join(cfg.PackageDir, "vim", "vimrc")
-	require.NoError(t, cfg.FS.WriteFile(ctx, sourcePath, []byte("test"), 0644))
-
-	// Create relative symlink
-	linkPath := filepath.Join(cfg.TargetDir, ".vimrc")
-	relTarget := "../" + filepath.Base(cfg.PackageDir) + "/vim/vimrc"
-	require.NoError(t, cfg.FS.Symlink(ctx, relTarget, linkPath))
-
-	manifestContent := []byte(`{
-		"version": "1.0",
-		"updated_at": "2024-01-01T00:00:00Z",
-		"packages": {
-			"vim": {"name": "vim", "installed_at": "2024-01-01T00:00:00Z", "link_count": 1, "links": [".vimrc"]}
-		},
-		"hashes": {}
-	}`)
-	manifestPath := filepath.Join(cfg.TargetDir, ".dot-manifest.json")
-	require.NoError(t, cfg.FS.WriteFile(ctx, manifestPath, manifestContent, 0644))
-
-	client, err := dot.NewClient(cfg)
-	require.NoError(t, err)
-
-	report, err := client.DoctorWithScan(ctx, dot.DefaultScanConfig())
-	require.NoError(t, err)
-
-	// Relative symlink with valid target should be healthy
-	assert.Equal(t, dot.HealthOK, report.OverallHealth)
-	assert.Equal(t, 0, report.Statistics.BrokenLinks)
-}
-
-func TestDoctor_PermissionError(t *testing.T) {
-	// This test would require mocking FS to return permission errors
-	// Skipping for now as our memfs doesn't simulate permissions
-	t.Skip("Requires permission simulation")
-}
-
-func TestList_CallsStatus(t *testing.T) {
-	cfg := testConfig(t)
-	ctx := context.Background()
-
-	client, err := dot.NewClient(cfg)
-	require.NoError(t, err)
-
-	// List delegates to Status
-	packages, err := client.List(ctx)
-	require.NoError(t, err)
-
-	status, err := client.Status(ctx)
-	require.NoError(t, err)
-
-	// Should return same data
-	assert.Equal(t, status.Packages, packages)
-}
-
-func TestAdopt_UpdatesManifest(t *testing.T) {
-	cfg := testConfig(t)
-	ctx := context.Background()
-
-	// Create file
-	require.NoError(t, cfg.FS.MkdirAll(ctx, cfg.TargetDir, 0755))
-	filePath := filepath.Join(cfg.TargetDir, ".bashrc")
-	require.NoError(t, cfg.FS.WriteFile(ctx, filePath, []byte("test"), 0644))
-
-	pkgDir := filepath.Join(cfg.PackageDir, "bash")
-	require.NoError(t, cfg.FS.MkdirAll(ctx, pkgDir, 0755))
-
-	client, err := dot.NewClient(cfg)
-	require.NoError(t, err)
-
-	// Adopt (will fail during execution due to preconditions, but tests Adopt path)
-	_ = client.Adopt(ctx, []string{".bashrc"}, "bash")
-	// Error expected due to executor preconditions, but Adopt code path is covered
-}
